migrerer tekst til long for fødselsnummer og aktørId
fjerner også utbetalingsreferanse-tabellen som ikke har vært i bruk på snart ett år.

+++ sykepenger-api/src/main/kotlin/no/nav/helse/spleis/dao/PersonDao.kt
@@ -9,10 +9,10 @@ import javax.sql.DataSource
 
 internal class PersonDao(private val dataSource: DataSource) {
     fun hentPerson(fødselsnummer: String) =
-        hentPerson(queryOf("SELECT data FROM person WHERE fnr = ? ORDER BY id DESC LIMIT 1", fødselsnummer))
+        hentPerson(queryOf("SELECT data FROM person WHERE fnr = ? ORDER BY id DESC LIMIT 1", fødselsnummer.toLong()))
 
     fun hentPersonAktørId(aktørId: String) =
-        hentPerson(queryOf("SELECT data FROM person WHERE aktor_id = ? ORDER BY id DESC LIMIT 1", aktørId))
+        hentPerson(queryOf("SELECT data FROM person WHERE aktor_id = ? ORDER BY id DESC LIMIT 1", aktørId.toLong()))
 
     private fun hentPerson(query: Query) =
         using(sessionOf(dataSource)) { session ->
