+++ sykepenger-mediators/src/main/kotlin/no/nav/helse/spleis/db/PersonPostgresRepository.kt
@@ -6,13 +6,12 @@ import kotliquery.sessionOf
 import kotliquery.using
 import no.nav.helse.serde.SerialisertPerson
 import no.nav.helse.spleis.PostgresProbe
-import java.util.*
 import javax.sql.DataSource
 
 internal class PersonPostgresRepository(private val dataSource: DataSource) : PersonRepository {
 
     override fun hentPerson(fødselsnummer: String) =
-        hentPerson(queryOf("SELECT data FROM person WHERE fnr = ? ORDER BY id DESC LIMIT 1", fødselsnummer))
+        hentPerson(queryOf("SELECT data FROM person WHERE fnr = ? ORDER BY id DESC LIMIT 1", fødselsnummer.toLong()))
 
     private fun hentPerson(query: Query) =
         using(sessionOf(dataSource)) { session ->
